setup new report generation scripts
